# TRAAN_Integration_Phase5/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH Disaster Response API",
    "description": "Disaster Response Coordination Platform \u2014 Backend API\n\n**SIH 2026** \u00b7 Component D \u2014 Backend Core\n\nThis Swagger UI is the live contract that Android, Dashboard, and AI/Optimizer teams build against. Do not change endpoint paths, request fields, or response fields without coordinating with all affected teams.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/incidents": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Incidents",
        "operationId": "list_incidents_api_v1_incidents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bbox"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedIncidentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Incident",
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "incidents"
        ],
        "summary": "Update Incident",
        "operationId": "update_incident_api_v1_incidents__incident_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/recommendations": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Recommendations",
        "operationId": "get_recommendations_api_v1_incidents__incident_id__recommendations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/dispatch": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Create Dispatch",
        "operationId": "create_dispatch_api_v1_incidents__incident_id__dispatch_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/refresh-summary": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Refresh Summary",
        "description": "Integration boundary stub for AI-based summary refresh.\nComponent E is responsible for the actual implementation.",
        "operationId": "refresh_summary_api_v1_incidents__incident_id__refresh_summary_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DispatchRequest": {
        "properties": {
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "resource_id",
          "quantity"
        ],
        "title": "DispatchRequest"
      },
      "DispatchResponse": {
        "properties": {
          "dispatch_id": {
            "type": "string",
            "title": "Dispatch Id"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "quantity_dispatched": {
            "type": "integer",
            "title": "Quantity Dispatched"
          },
          "dispatched_by": {
            "type": "string",
            "title": "Dispatched By"
          },
          "dispatched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Dispatched At"
          },
          "eta_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Minutes"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "dispatch_id",
          "incident_id",
          "resource_id",
          "quantity_dispatched",
          "dispatched_by",
          "dispatched_at",
          "status"
        ],
        "title": "DispatchResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentFlags": {
        "properties": {
          "medical_emergency": {
            "type": "boolean",
            "title": "Medical Emergency"
          },
          "trapped": {
            "type": "boolean",
            "title": "Trapped"
          },
          "elderly_or_children": {
            "type": "boolean",
            "title": "Elderly Or Children"
          },
          "structural_damage": {
            "type": "boolean",
            "title": "Structural Damage"
          }
        },
        "type": "object",
        "required": [
          "medical_emergency",
          "trapped",
          "elderly_or_children",
          "structural_damage"
        ],
        "title": "IncidentFlags"
      },
      "IncidentResponse": {
        "properties": {
          "area_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Name"
          },
          "emergency_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Emergency Types"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "report_count": {
            "type": "integer",
            "title": "Report Count"
          },
          "estimated_people_affected": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated People Affected"
          },
          "flags": {
            "$ref": "#/components/schemas/IncidentFlags"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "recommended_resources": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Recommendation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Resources"
          },
          "assigned_resources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Resources"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "cluster_id": {
            "type": "string",
            "title": "Cluster Id"
          },
          "source_sos_reports": {
            "items": {},
            "type": "array",
            "title": "Source Sos Reports"
          },
          "location": {
            "$ref": "#/components/schemas/app__models__sos__Location"
          },
          "first_reported_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Reported At"
          },
          "last_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated At"
          }
        },
        "type": "object",
        "required": [
          "emergency_types",
          "severity",
          "report_count",
          "flags",
          "status",
          "incident_id",
          "cluster_id",
          "source_sos_reports",
          "location",
          "first_reported_at",
          "last_updated_at"
        ],
        "title": "IncidentResponse"
      },
      "IncidentStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "IncidentStatusUpdate"
      },
      "PaginatedIncidentsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedIncidentsResponse"
      },
      "Recommendation": {
        "properties": {
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning"
          }
        },
        "type": "object",
        "required": [
          "resource_type",
          "quantity",
          "reasoning"
        ],
        "title": "Recommendation"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__models__sos__Location": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng"
          },
          "accuracy_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy M"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "title": "Location"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}